--- src/app/@authModal/(.)login/login.module.css
@@ -1,82 +1,0 @@
-.rightSideArea {
-  width: 570px;
-  padding: 70px 80px;
-
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-
-  background-color: var(--color-white);
-}
-
-.loginForm {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.loginButton {
-  margin-top: 24px;
-  padding: 14px;
-
-  border: none;
-  border-radius: 8px;
-  background-color: var(--color-primary-500);
-
-  font-size: 15px;
-  font-weight: 600;
-  color: var(--color-white);
-
-  cursor: pointer;
-}
-
-.signupArea {
-  margin-top: 18px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.text {
-  font-size: 18px;
-  font-weight: 500;
-  color: var(--color-gray-700);
-}
-
-.bold {
-  font-weight: 700;
-}
-
-.signupArea button {
-  padding: 14px;
-
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  justify-content: center;
-
-  border-radius: 8px;
-
-  font-size: 15px;
-  font-weight: 600;
-
-  cursor: pointer;
-}
-
-.kakao {
-  border: none;
-  background-color: #fee500;
-  color: var(--color-gray-700);
-}
-
-.naver {
-  border: none;
-  background-color: #03c75a;
-  color: var(--color-white);
-}
-
-.google {
-  border: 1px solid var(--color-gray-300);
-  background-color: var(--color-white);
-  color: var(--color-gray-700);
-}
