+++ src/components/common/button/Button.module.css
@@ -1,5 +1,5 @@
 .button {
-  padding: 14px 20px;
+  padding: 15px 20px;
   font-size: 15px;
   font-weight: 600;
   border: 0 none;
