+++ src/components/common/button/linkButton.module.css
@@ -1,5 +1,5 @@
 .button {
-  height: 48px;
+  padding: 15px 20px;
 
   display: flex;
   justify-content: center;
@@ -11,4 +11,10 @@
   font-weight: 600;
   color: var(--color-primary-500);
   text-decoration: none;
+}
+
+.fill {
+  border: none;
+  color: var(--color-white);
+  background-color: var(--color-primary-500);
 }
