--- src/hooks/1.txt
@@ -1,1 +1,0 @@
-use~와 같은 훅 폴더
