+++ src/hooks/useAuthInputState.tsx
@@ -1,0 +1,1 @@
+const useAuthInputState = () => {}
